The bot structure has been redesigned. Again
