+++ requirements/testing.txt
@@ -50,6 +50,7 @@ pymemcache==4.0.0
 pyproject_hooks==1.1.0
 pytest==8.3.2
 pytest-asyncio==0.23.8
+python-decouple==3.8
 python-dotenv==1.0.1
 pytz==2024.1
 rapidfuzz==3.9.6
